feat(DTFS2-7298): rename function
