+++ trade-finance-manager-ui/server/controllers/case/index.js
@@ -2,7 +2,7 @@ const { format, fromUnixTime } = require('date-fns');
 const { isEmpty } = require('lodash');
 const { AMENDMENT_STATUS, isTfmFacilityEndDateFeatureFlagEnabled } = require('@ukef/dtfs2-common');
 const api = require('../../api');
-const { getTask, showAmendmentButton, ukefDecisionRejected, isDealCancellationEnabled } = require('../helpers');
+const { getTask, showAmendmentButton, ukefDecisionRejected, canDealBeCancelled } = require('../helpers');
 const { formattedNumber } = require('../../helpers/number');
 const mapAssignToSelectOptions = require('../../helpers/map-assign-to-select-options');
 const CONSTANTS = require('../../constants');
@@ -45,7 +45,7 @@ const getCaseDeal = async (req, res) => {
 
   const { submissionType } = deal.dealSnapshot;
 
-  const dealCancellationIsEnabled = isDealCancellationEnabled(submissionType, user);
+  const dealCancellationIsEnabled = canDealBeCancelled(submissionType, user);
   let hasDraftCancellation = false;
 
   if (dealCancellationIsEnabled) {
