+++ trade-finance-manager-ui/server/controllers/helpers/can-deal-be-cancelled.helper.ff-test.ts
@@ -1,5 +1,5 @@
 import { DEAL_SUBMISSION_TYPE, TEAM_IDS } from '@ukef/dtfs2-common';
-import { isDealCancellationEnabled } from './deal-cancellation-enabled.helper';
+import { canDealBeCancelled } from './can-deal-be-cancelled.helper';
 import { TfmSessionUser } from '../../types/tfm-session-user';
 
 const pimUser = { teams: [TEAM_IDS.PIM] } as TfmSessionUser;
@@ -7,23 +7,23 @@ const nonPimUser = { teams: [TEAM_IDS.UNDERWRITERS] } as TfmSessionUser;
 
 const { AIN, MIN, MIA } = DEAL_SUBMISSION_TYPE;
 
-describe('dealCancellationEnabled', () => {
+describe('canDealBeCancelled', () => {
   describe('when `FF_TFM_FACILITY_END_DATE_ENABLED` is set to true', () => {
     describe('when user is a PIM user', () => {
       it('should return false if the deal type is MIA', () => {
-        const result = isDealCancellationEnabled(MIA, pimUser);
+        const result = canDealBeCancelled(MIA, pimUser);
 
         expect(result).toEqual(false);
       });
 
       it('should return true if the deal type is MIN', () => {
-        const result = isDealCancellationEnabled(MIN, pimUser);
+        const result = canDealBeCancelled(MIN, pimUser);
 
         expect(result).toEqual(true);
       });
 
       it('should return true if the deal type is AIN', () => {
-        const result = isDealCancellationEnabled(AIN, pimUser);
+        const result = canDealBeCancelled(AIN, pimUser);
 
         expect(result).toEqual(true);
       });
@@ -31,7 +31,7 @@ describe('dealCancellationEnabled', () => {
 
     describe('when user is not a PIM user', () => {
       it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabled(type, nonPimUser);
+        const result = canDealBeCancelled(type, nonPimUser);
 
         expect(result).toEqual(false);
       });
