+++ trade-finance-manager-ui/server/controllers/helpers/can-deal-be-cancelled.helper.test.ts
@@ -1,5 +1,5 @@
 import { DEAL_SUBMISSION_TYPE, TEAM_IDS } from '@ukef/dtfs2-common';
-import { canSubmissionTypeBeCancelled, isDealCancellationEnabled } from './deal-cancellation-enabled.helper';
+import { canSubmissionTypeBeCancelled, canDealBeCancelled } from './can-deal-be-cancelled.helper';
 import { TfmSessionUser } from '../../types/tfm-session-user';
 
 const pimUser = { teams: [TEAM_IDS.PIM] } as TfmSessionUser;
@@ -7,11 +7,11 @@ const nonPimUser = { teams: [TEAM_IDS.UNDERWRITERS] } as TfmSessionUser;
 
 const { AIN, MIN, MIA } = DEAL_SUBMISSION_TYPE;
 
-describe('dealCancellationEnabled', () => {
+describe('canDealBeCancelled', () => {
   describe('when `FF_TFM_FACILITY_END_DATE_ENABLED` is set to false', () => {
     describe('when user is a PIM user', () => {
       it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabled(type, pimUser);
+        const result = canDealBeCancelled(type, pimUser);
 
         expect(result).toEqual(false);
       });
@@ -19,7 +19,7 @@ describe('dealCancellationEnabled', () => {
 
     describe('when user is not a PIM user', () => {
       it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabled(type, nonPimUser);
+        const result = canDealBeCancelled(type, nonPimUser);
 
         expect(result).toEqual(false);
       });
